pyproject.toml
[project]
name = "agente-gastos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
    "uvicorn>=0.35.0",
]
